Add test that single attendee endpoint returns error if ET+ not loaded
